casatables: handle WEIGHT correctly in spwglue
In the Python version I had coded that WEIGHT should be the same for every
input spw, but that isn't the case in my updated pipeline that applies proper
weights before the glue stage. There was just a bug in the Python code that
made it so we only took the first WEIGHT cell value encountered instead.

Here we average all of the WEIGHT vectors, which I think is the right thing to
do.