casatables: fix Table::column_names()
I'm now finding that we're no longer able to get away with the lifetime
shenanigans that we were pulling in Table::column_names() before. As far as I
can tell, we need to use C++ => Rust callbacks to be absolutely sure that we
can safely get string data from the C++ layer to Rust. So, let's go ahead and
do it. Fortunately I've figured out a boilerplate that works for the "table
keywords" interface, so this is more tedious than anything else.