Once more unto the string breach
As reported in #133, there are *still* problems with passing strings
back and forth. I think this update should fix them for realsies. If we
want to get strings out of C++ into Rust, the only (correct) way to do
so reliably is to invoke a Rust callback from the C++ code that copies
the data. There's just no way to otherwise ensure that C++s "lifetimes"
agree with Rust's, and there's always a possibility that we'll need to
copy the data anyway because we need to make sure that they are valid
UTF8.

This change updates the C++ glue layer to ensure that "unbridge"
operations always invoke a callback, and then updates the Rust
appropriately. As part of this, I've somewhat tidied up the callback
infrastructure.

Closes #133 (I hope).