Make ndarray and num-complex versions flexible.
Not all crates are quick to update their dependencies, and this can mean
that mixing the results out of rubbl with other crates becomes a
headache, even if e.g. the ndarray array to be used across libraries is
represented the exact same way, because different libraries use
different versions of ndarray. By allowing these dependencies to be
flexible in their versions, hopefully, a single version of ndarray or
num-complex can be used to resolve multiple crate dependencies. In this
way, ndarray and num-complex are particularly painful, because they
mainly provide data types and these have not fundamentally changed over
the last few minor version bumps.

ndarray 0.8 and num-complex 0.1 are the minimum versions of these crates
that allowed compilation to succeed and tests to pass.