[pkgw/rubbl#170] pub use approx in core
From @pkgw: "My only comment would be that we should pub use the approx crate in core/src/lib.rs so that downstream code has a way to refer to the dep in code. Without this, I believe that you can get weird conflicts if your app ends up depending on multiple versions of approx — you can have two traits named AbsDiffEq that the compiler considers to be different. By pub using the crate, you at least give people the ability to make sure that they're talking about the intended version. Also, that pub use would be the place to add a comment in the source code explaining why we depend on the approx crate even though we don't actually use it in Rubbl (yet)."