Specify minimum versions for all dependencies.
This is important because some dependency versions don't actually work,
e.g. thiserror = "^1.0" allows 1.0.0, but actually >=1.0.7 is what is
required for rubbl to compile. This commit also removes all of the "^"
versions because (IMHO) it's clearer to use the major.minor.patch layout
and it does the same thing.

This commit removes version ranges on ndarray and num-complex (I
introduced these in the first place, thinking they were a good idea).
Why are version ranges bad? At the time of writing, 0.15.6 is the latest
version of ndarray. If my code depends on ndarray 0.15.0 and rubbl
depends on ndarray >= 0.8, cargo will determine that ndarray 0.15.6 is
fit for purpose. But, if tomorrow ndarray releases 0.16.0, then cargo
will inspect these requirements and determine that 0.16.0 is best for
rubbl and 0.15.6 is best for my code. Now I depend on two versions of
ndarray, and I can't interact my ndarray code with rubbl's ndarray code.
This is a huge pain, so it's (IHMO) simpler to have libraries (like
rubbl) depend on their oldest semver compatible versions, i.e. today
that is 0.15.0. If/when ndarray 0.16.0 is released, then others
depending on ndarray 0.15.0 and rubbl won't have their code break, and
rubbl should be eventually updated to support ndarray 0.16.0, allowing
rubbl consumers to also update.

clap is a bit of an exception; 4.0.0 probably works, but clap is mostly
intended for binary usage, not library usage, so I think it makes the
most sense to enforce a modern version of clap with lots of bug fixes.