Clean up dependencies.
Make anyhow, clap and termcolor optional dependencies in rubbl_core;
they are now activated by the "notifications" feature. This means that
downstream consumers of e.g. rubbl_casacore don't pull these
dependencies in which aren't actually used in the library.

Remove approx as a dependency. This was only in the hope of reducing
friction between date-type-exporting crates like num-complex, but, (now
that we all understand this a bit better,) this doesn't do anything
here.