Remove `casatables::NotScalarColumnError`
In the update to thiserror/anyhow, this error type was brought into the
`TableError` type. While similar variants like
`TableError::UnexpectedDataType` continued to refer to the separate
`UnexpectedDataTypeError`, the `TableError::NotScalarColumnError`
variant does not, making this type superfluous.

Our error API is now a little bit wonky:

- For naming consistency, `TableError::NotScalarColumnError` should
  become `TableError::NotScalarColumn`
- The `UnexpectedDataTypeError` type should probably also go away

But those are slightly bigger API changes, so I'm avoiding them for now.