Update to ndarray 0.16 and improve casatables uninitialized data
Here we finally fix the warning about uninitialized data in our
casatables array buffers. We need to tweak the CasaDataType trait a bit,
but no one external should have been implementing it themselves, so that
shouldn't be an issue. It took me some time to get the trait
specifications right but in the end it's not too complicated.