casatables: mark TableError as inexhaustive and add a UserMessage variant
I ran into an issue with `Table::for_each_row` where I wanted to issue
an application-specific error message. The return type of the callback
function has to be a TableError, though, which wasn't providing an
option to do this. Add one.

Also mark the enum as `non_exhaustive`, which is technically an API
break now but makes us more extensible in the future.

A sensible extension would be to add another variant for a generic boxed
error. That might affect sendability or something in a way that we would
want to think about.